--- a/dpc-portal/app/jobs/verify_provider_organization_job.rb
+++ b/dpc-portal/app/jobs/verify_provider_organization_job.rb
@@ -13,6 +13,7 @@ class VerifyProviderOrganizationJob < ApplicationJob
     rescue AoException => e
       handle_error(org, e.message)
     end
+    puts "Gateway calls: #{service.cpi_api_gw_client.counter}"
   end
 
   def handle_error(org, message)