--- a/dpc-portal/app/services/cpi_api_gateway_client.rb
+++ b/dpc-portal/app/services/cpi_api_gateway_client.rb
@@ -5,6 +5,7 @@ require 'oauth2'
 # A client for requests to the CPI API Gateway
 class CpiApiGatewayClient
   attr_accessor :access
+  attr_reader :counter
 
   def initialize
     env = ENV.fetch('ENV', nil)
@@ -13,6 +14,8 @@ class CpiApiGatewayClient
     cms_idm_url = ENV.fetch('CMS_IDM_OAUTH_URL', nil)
     @cpi_api_gateway_url = ENV.fetch('CPI_API_GW_BASE_URL', nil)
     @cpi_api_gateway_url += '/' unless @cpi_api_gateway_url.end_with?('/')
+
+    @counter = 0
     @client = OAuth2::Client.new(client_id, client_secret,
                                  site: cms_idm_url,
                                  token_url: '/oauth2/aus2151jb0hszrbLU297/v1/token',
@@ -24,6 +27,7 @@ class CpiApiGatewayClient
 
   # fetch data about an organization, including enrollment_id
   def fetch_enrollment(npi)
+    @counter += 1
     body = { providerID: { npi: npi.to_s } }.to_json
     response = request_client.post("#{@cpi_api_gateway_url}api/1.0/ppr/providers/enrollments",
                                    headers: { 'Content-Type': 'application/json' },
@@ -33,6 +37,7 @@ class CpiApiGatewayClient
 
   # fetch a list of roles, roughly corresponding to associated individuals
   def fetch_enrollment_roles(enrollment_id)
+    @counter += 1
     response = request_client.get("#{@cpi_api_gateway_url}api/1.0/ppr/providers/enrollments/#{enrollment_id}/roles",
                                   headers: { 'Content-Type': 'application/json' })
     response.parsed
@@ -40,6 +45,7 @@ class CpiApiGatewayClient
 
   # fetch info about the authorized official, including a list of med sanctions
   def fetch_med_sanctions_and_waivers_by_ssn(ssn)
+    @counter += 1
     body = {
       providerID: {
         providerType: 'ind',
@@ -57,6 +63,7 @@ class CpiApiGatewayClient
 
   # fetch info about the organization, including a list of med sanctions
   def fetch_med_sanctions_and_waivers_by_org_npi(npi)
+    @counter += 1
     body = {
       providerID: {
         providerType: 'org',