## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3848

## 🛠 Changes

- Rake task added to test 500 organizations, verify expected outcomes
- CPI API GW updated to count calls made
- SyncOrganizationJob turned off as unnecessary to the test
- Fake gateway updated to accommodate test

## ℹ️ Context for reviewers

Test to make sure meets minimum performance needs

## ✅ Acceptance Validation

500 organizations created:
  - 10 organizations fail med sanction check
  - 15 organizations have no enrollments
  - 20 organizations no longer have a valid AO
  - 5 organizations have an AO that fails med sanction check
  - 50 organizations do not have an AO
  - 50 organizations are set up so that one AO manages both of them
  - 50 organizations are set up so that one AO manages four of them
Results of test:
 - All expected results of conditions are met
 - The VerifyAoJob took 3.1 seconds
 - The VerifyAoJob made 1300 calls to the CPI API Gateway
 - The VerifyProviderOrganizationJob took 1.4 seconds
 - The VerifyProviderOrganizationJob made 950 calls to the CPI API Gateway

Assuming we can achieve 25 calls per second, checking 500 organizations should take 95 seconds.
(1300 + 950)/25 -> 90 seconds, plus 3.1 + 1.4 -> 95 seconds

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security controls.
- [ ] This PR stores or transmits data that was not stored or transmitted before.
- [ ] This PR requires additional review of its security implications for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username: StewGoin) as a reviewer and do not merge this PR without his approval.